requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10.1
djangorestframework==3.4.6
enum34==1.1.6
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
wcwidth==0.1.7
